requirements.txt
mysql-connector-python>=8.0.0
Pillow>=9.0.0
